aws-ssm/get_parameters_and_write_file/requirements.txt
boto3==1.26.56
progress==1.6
python-dotenv==0.21.1
